Extracting more bit data logic into `bitData*.js`
